Fixing bug with global/default XML jobs
The original Magento code would look under crontab/jobs and default/crontab/jobs
to find a job's config. The Aoe_Scheduler code treated these are two separate types
of jobs. The problem with this is that the collection that was collecting all the
jobs would not let you add two entries for the same code so the default/crontab/jobs
entries that were meant to be overrides of the crontab/jobs entries were being ignored.

This update merges crontab/jobs and default/crontab/jobs into a single job type of 'xml'.
The default/crontab/jobs values will override the crontab/jobs values.