Make a difference between missed and skipped jobs
Missed jobs are missed because they did not run in time. Another job postponed them for too long.
Skipped jobs are duplicate jobs of a job that is scheduled to run now. They were postponed, but are still within the time limit to run. They are not missed, but skipped because another instance makes them unnecessary.